requirements.txt
beautifulsoup4==4.12.3
comfyui==0.0.1
langchain_chroma==0.1.3
langchain_community==0.2.15
langchain_core==0.2.37
langchain_huggingface==0.0.3
langchain_openai==0.1.23
markdown_to_json==2.1.1
openai==1.43.0
pandas==2.2.2
PyYAML==6.0.1
PyYAML==6.0.2
websocket_client==1.8.0
